Fetch paginated accounts when completing session.
