+++ financial-connections/src/main/java/com/stripe/android/financialconnections/domain/FetchFinancialConnectionsSession.kt
@@ -1,14 +1,11 @@
 package com.stripe.android.financialconnections.domain
 
-import com.stripe.android.financialconnections.FinancialConnectionsSheetViewModel
-import com.stripe.android.financialconnections.model.FinancialConnectionsAccount
-import com.stripe.android.financialconnections.model.FinancialConnectionsAccountList
 import com.stripe.android.financialconnections.model.FinancialConnectionsSession
-import com.stripe.android.financialconnections.model.GetFinancialConnectionsAcccountsParams
 import com.stripe.android.financialconnections.repository.FinancialConnectionsRepository
 import javax.inject.Inject
 
 internal class FetchFinancialConnectionsSession @Inject constructor(
+    private val fetchPaginatedAccountsForSession: FetchPaginatedAccountsForSession,
     private val financialConnectionsRepository: FinancialConnectionsRepository
 ) {
 
@@ -21,36 +18,6 @@ internal class FetchFinancialConnectionsSession @Inject constructor(
      */
     suspend operator fun invoke(clientSecret: String): FinancialConnectionsSession {
         val session = financialConnectionsRepository.getFinancialConnectionsSession(clientSecret)
-        if (session.accounts.hasMore) {
-            val accounts = mutableListOf<FinancialConnectionsAccount>()
-            accounts.addAll(session.accounts.data)
-
-            var nextAccountList = financialConnectionsRepository.getFinancialConnectionsAccounts(
-                GetFinancialConnectionsAcccountsParams(clientSecret, accounts.last().id)
-            )
-            accounts.addAll(nextAccountList.data)
-
-            while (nextAccountList.hasMore && accounts.size < FinancialConnectionsSheetViewModel.MAX_ACCOUNTS) {
-                nextAccountList = financialConnectionsRepository.getFinancialConnectionsAccounts(
-                    GetFinancialConnectionsAcccountsParams(clientSecret, accounts.last().id)
-                )
-                accounts.addAll(nextAccountList.data)
-            }
-
-            return FinancialConnectionsSession(
-                id = session.id,
-                clientSecret = session.clientSecret,
-                accountsNew = FinancialConnectionsAccountList(
-                    data = accounts,
-                    hasMore = nextAccountList.hasMore,
-                    url = nextAccountList.url,
-                    count = accounts.size,
-                    totalCount = nextAccountList.totalCount
-                ),
-                bankAccountToken = null, // Token should not be exposed on regular sessions.
-                livemode = session.livemode
-            )
-        }
-        return session
+        return fetchPaginatedAccountsForSession(session)
     }
 }
